+++ skoolkit/snapmod.py
@@ -37,8 +37,8 @@ def main(args):
     group = parser.add_argument_group('Options')
     group.add_argument('-m', '--move', dest='moves', metavar='src,size,dest', action='append', default=[],
                        help='Move a block of bytes of the given size from src to dest. This option may be used multiple times.')
-    group.add_argument('-p', '--poke', dest='pokes', metavar='a[-b[-c]],[^+]v', action='append', default=[],
-                       help="POKE N,v for N in {a, a+c, a+2c..., b}. "
+    group.add_argument('-p', '--poke', dest='pokes', metavar='[p:]a[-b[-c]],[^+]v', action='append', default=[],
+                       help="POKE N,v in RAM bank p for N in {a, a+c, a+2c..., b}. "
                             "Prefix 'v' with '^' to perform an XOR operation, or '+' to perform an ADD operation. "
                             "This option may be used multiple times.")
     group.add_argument('-r', '--reg', dest='reg', metavar='name=value', action='append', default=[],