+++ skoolkit/trace.py
@@ -300,8 +300,8 @@ def main(args):
                        help="Don't execute interrupt routines.")
     group.add_argument('-o', '--org', metavar='ADDR', type=integer,
                        help='Specify the origin address of a binary (raw memory) file (default: 65536 - length).')
-    group.add_argument('-p', '--poke', dest='pokes', metavar='a[-b[-c]],[^+]v', action='append', default=[],
-                       help="POKE N,v for N in {a, a+c, a+2c..., b}. "
+    group.add_argument('-p', '--poke', dest='pokes', metavar='[p:]a[-b[-c]],[^+]v', action='append', default=[],
+                       help="POKE N,v in RAM bank p for N in {a, a+c, a+2c..., b}. "
                             "Prefix 'v' with '^' to perform an XOR operation, or '+' to perform an ADD operation. "
                             "This option may be used multiple times.")
     group.add_argument('-r', '--reg', metavar='name=value', action='append', default=[],